Remove `lib` folder from autoload path
- Load extensions explicitly
