+++ app/controllers/epp_controller.rb
@@ -1,5 +1,4 @@
 class EppController < ApplicationController
-  include Iptable
   layout false
   protect_from_forgery with: :null_session
   skip_before_action :verify_authenticity_token
@@ -406,5 +405,21 @@ class EppController < ApplicationController
   def session_timeout_reached?
     timeout = 5.minutes
     epp_session.updated_at < (Time.zone.now - timeout)
+  end
+
+  def counter_update(registrar_code, ip)
+    counter_proc = "/proc/net/xt_recent/#{registrar_code}"
+
+    begin
+      File.open(counter_proc, 'a') do |f|
+        f.puts "+#{ip}"
+      end
+    rescue Errno::ENOENT => e
+      logger.error "IPTABLES COUNTER UPDATE: cannot open #{counter_proc}: #{e}"
+    rescue Errno::EACCES => e
+      logger.error "IPTABLES COUNTER UPDATE: no permission #{counter_proc}: #{e}"
+    rescue IOError => e
+      logger.error "IPTABLES COUNTER UPDATE: cannot write #{ip} to #{counter_proc}: #{e}"
+    end
   end
 end
