+++ config/initializers/devise.rb
@@ -1,3 +1,5 @@
+require 'devise_custom_failure'
+
 # Use this hook to configure devise mailer, warden hooks and so forth.
 # Many of these configuration options can be set straight in your model.
 Devise.setup do |config|
