+++ config/initializers/ext.rb
@@ -1,0 +1,4 @@
+require 'core_ext/array'
+require 'gem_ext/builder'
+require 'gem_ext/i18n'
+require 'gem_ext/paper_trail'
