--- config/initializers/load_class_extensions.rb
@@ -1,1 +1,0 @@
-Dir[File.join(Rails.root, 'lib', 'ext', '*.rb')].each { |x| require x }
