+++ lib/core_ext/array.rb
