+++ lib/gem_ext/i18n.rb
@@ -1,5 +1,5 @@
 # Don't raise error when nil
-# http://stackoverflow.com/questions/9467034/rails-i18n-how-to-handle-case-of-a-nil-date-being-passed-ie-lnil 
+# http://stackoverflow.com/questions/9467034/rails-i18n-how-to-handle-case-of-a-nil-date-being-passed-ie-lnil
 module I18n
   class << self
     alias_method :original_localize, :localize
