+++ lib/gem_ext/paper_trail.rb
